Refactor code
