+++ config.js
@@ -2,5 +2,10 @@ module.exports = {
     /*
      * Jitsi Meet domain
      */
-    JITSI_MEET_DOMAIN: 'meet.jit.si'
+    JITSI_MEET_DOMAIN: 'meet.jit.si',
+
+    /*
+     * Port number to listen to commands
+     */
+    PORT: process.env.PORT || 8000
 };
